--- src/entities/lab-main/index.ts
@@ -1,2 +1,0 @@
-export { LabLinkBtn } from './ui/lab-link-btn'
-export { LabDownloadData } from './ui/lab-download-data'
