+++ src/entities/lab/index.ts
@@ -1,2 +1,4 @@
 export { InviteLab } from './ui/invite-lab'
 export { LabUser } from './ui/lab-user'
+export { LabLinkItem } from './ui/lab-link-item'
+export { LabIntro } from './ui/lab-intro'
