+++ src/entities/lab/ui/lab-intro.tsx
