+++ src/widgets/lab-widgets/index.ts
@@ -1,3 +1,2 @@
 export { LabLink } from './lab-link'
 export { LabMember } from './lab-member'
-export { LabIntro } from './lab-intro'
