+++ src/widgets/lab-widgets/lab-link.tsx
@@ -1,6 +1,6 @@
 import { Image } from 'next/dist/client/image-component'
 import { useState } from 'react'
-import { LabLinkBtn } from '@/entities/lab-main'
+import { LabLinkItem } from '@/entities/lab'
 import { LabLinkModal } from '@/widgets/modal'
 
 export const LabLink = () => {
@@ -20,10 +20,10 @@ export const LabLink = () => {
         />
       </div>
       <div className={'flex flex-wrap gap-[6px]'}>
-        <LabLinkBtn href={'#'} name={'연구실 홈페이지'} />
-        <LabLinkBtn href={'#'} name={'엑셀'} />
-        <LabLinkBtn href={'#'} name={'이것저것'} />
-        <LabLinkBtn href={'#'} name={'링크'} />
+        <LabLinkItem href={'#'} name={'연구실 홈페이지'} />
+        <LabLinkItem href={'#'} name={'엑셀'} />
+        <LabLinkItem href={'#'} name={'이것저것'} />
+        <LabLinkItem href={'#'} name={'링크'} />
       </div>
       <LabLinkModal isOpen={isOpen} onClose={() => setIsOpen(false)} />
     </div>
